Implement TCP listener
The program can now listen on a TCP port for incoming connections. Once a client is connected, it handles it much like an outbound connection, by sending from stdin and printing out to stdout.

It actually listens on the IPv4 and IPv6 wildcard addresses and supports serially receiving connections on both of them.

Now that the command line args are becoming more numerous, this change also switches to using the `clap` command line argument parser. Hey, that's what that stands for!