Implement UDP listener
This change implements the UDP listener. It can listen on multiple UDP sockets at the same time and prints out traffic to stdout from all of them. The first remote peer that sent traffic will also be selected as the destination for stdin data. It prints out each unique remote peer as it is encountered.