Allow specifying source address as well as port
All listening and outbound connection functionality now allows specifying the `-S` command line argument to specify the local address to bind to.

For outbound connections, if the specified address doesn't match the address family of the destination, it results in an error. If specifying a symbolic hostname, such as localhost, all resolved addresses will be tried until the first compatible one is found.

With this change, all logic for determining the local address and port to bind to is consolidated in one function call from one call site.