Add -4 and -6 to select address family
This adds the ability to filter to a specific address family (IPv4 or IPv6) by passing `-4` or `-6`. At most one may be specified at a time. If neither is specified, the program will try both (depends on the mode being used).

Also fixes a bug where UDP outgoing connections were hitting an assert.