Rewrite stdin read path using Stream/Sink
From referencing the tokio connect sample, this changes mainly the TCP send path to read from stdin and pass to the TCP socket using FramedRead, BytesCodec, and send_all.

This continues to not use tokio's async Stdin functionality because it doesn't offer cancelation on the stdin read. So if the socket is closed instead of stdin, the program will continue to run until the user hits enter to produce the next data in stdin and give an opportunity to bail before the next stdin read.

Instead, we use a thread that reads from stdin and sends to an mpsc channel. The thread can be exited at any time, giving a better interactive stdin experience.

The stdin reader thread is now responsible for sending only the right amount of data that the user had requested, usually from the `--sb` parameter.