Send interactive console characters immediately instead of on Enter
Previously, the program used Rust's stdin helper, which only reads characters in when the user presses Enter. Now by default for an interactive terminal, it uses the `console` crate to read characters as they're typed. When a non-interactive terminal, such as piped from a file, it uses the normal stdin reader.

The user can also choose to disable "character mode" with the `-c` command line flag.