Allow setting Time-to-Live for packets
This change adds a new option, `--ttl` to set the time-to-live for a packet, the number of IP hops that a packet can survive before it's discarded.

In anticipation of exposing more sockopts in the future, this change adds a new way of setting this, even though this particular option is exposed by Tokio. Now we have a way of going from a Tokio socket to the underlying `socket2::Socket`, which exposes basically all the sockopts we'd ever want.

To do this we have to commit slight crimes by converting to/from a RawFd/RawSocket and introduce Windows- and Unix-specific code, because the methods differ just slightly.

Now more options are settable in a single function, including the send buffer and broadcast capability. Some of these have to get filtered out based on UDP/TCP or IPv4/IPv6 so as not to cause runtime errors.