Add multicast support
This change adds the ability for outbound UDP connections to join a multicast V4 or V6 group by specifying the `--mc` arg.

Along with this are two other multicast options:
- `--ttl` now applies the multicast TTL rather than unicast TTL if multicast is enabled
- `--mc_no_loop` disables multicast traffic being routed back to the local machine

When a user doesn't specify a source port when using multicast, the local port is chosen to be the same as the outbound port, which is typically what the user expects.

To support this, `configure_socket_options` was changed to `with_socket_from_socket2`, which executes a closure and passes in the socket2::Socket. This lets us add a second helper function for joining a multicast group.

The V6 multicast join is a pain because it requires specifying the local interface index (rather than a local address). So I had to bring in a new crate, default_net, which enumerates local interfaces and lets me translate from specified local address to an interface index.