Add "echo" mode
This adds a new mode of handling input, where input received from the network is echoed back to the network. You can enable this mode by using the `-m echo` command line option. This also replaces the `-c` option with `-m stdin-nochar` or `-m snc`.

To achieve this, we just create an mpsc channel and connect the "receive from network" part to one end and the "send to network" part to the other end. Now the program understands local IO in terms of a "data from local" `Stream` and a "data to local" `Sink`. This should let us add more input modes in the future, so long as we can create a stream and sink for it.