Allow specifying output mode
User can now pass the `-o` argument to choose the output mode, i.e. where incoming data from the network is printed out, if anywhere. In particular, the user can now specify output mode `none` to do no local printing even when input comes from stdin, and output mode `stdout` to print locally even when the input mode is to echo network traffic back.

Rename `-m` to `-i` because there are now two modes that can be specified. `-i` used to be an alias for `--ttl` to match ping.exe, but that was an arbitrary choice, and I think the mirror for `-o` is more useful.