Allow generating random data
The user can now specify a new input mode `-i random` or `-i r` for short that has the program generate random data instead of taking from stdin. The user can specify `--rsizemin` and `--rsizemax` to choose the range of sizes of sends (mostly impacts UDP). The user can specify `--rvals` to choose between `binary` (`b`) or `ascii` (`a`) data to send.

To implement this, we now have a new object called `RandBytesIter` that produces rand data. This is wrapped in a stream and used the same as other streams for I/O.

Also fixes a bug that made all outbound connections not work due to command line arg parsing. Oops.