Let redirect output to file actually complete
When redirecting stdout to a file, stdin would start out at EOF and would stop the program immediately since it was reached. This is of course not ideal; the desired behavior is to allow stdin to do nothing and instead just wait for stdout to be done.

This change fixes that by introducing a new `InputMode::Null` with command line value `none`. The program automatically selects this mode when redirecting stdout and not redirecting stdin.

Fixes a bug where output mode `none` was not being respected when choosing input mode `random`.

Remove auto-detection of output mode. In all cases, it defaults to stdout, and the user can override it to none if they choose.

As a cleanup item, create friendly types for the input stream and output sink so we can stop writing ridiculous long types in several places.