Support broker mode and multiple inbound clients
This massive change adds a new feature and also rewrites a lot of the guts of the program. The new feature is called "broker" mode, which modern netcat supports. It allows in listening mode to accept multiple incoming connections and forward data between all of them.

This can be used for firewall traversal scenarios, for example to establish a connection between two endpoints that are both behind firewalls, instead a broker could be run on a public IP somewhere and both endpoints could connect to it instead.

To start with, the program now supports multiple inbound connections. By default, they only send and receive traffic from local IO, but by passing `--broker`, the program will also forward any traffic from one socket to any other connected peers. To specify a limit of number of concurrent connections, use the `-m` parameter.

I also added a `-v` parameter to help with debugging the routing flow. It turns on per-packet debug prints.

To support this change, most of the socket and input routing in the program had to be rewritten. Now all sockets participating in the program register themselves with a router, which keeps track of their remote peer addresses. Each one is represented by a stream of data from the network and a sink to forward data back to the network. Local IO is also represented by a stream and sink. The routing code examines a stream of all traffic tagged with its source address and decides which remote peers to forward each data buffer to.

It sounds simple when I put it that way, but there was *a lot* of convincing the Rust type system how to convert between various streams' and sinks' data types.

The echo channel functionality was built straight into the routing logic, so the old code was just deleted. The UDP and TCP paths are now very similar and follow similar patterns, though UDP's lack of a socket to represent each remote peer made it just different enough to prevent a lot of code reuse.

UDP traffic is now handled much more by a library object, `UdpFramed`, which gives a stream and sink much like a `TcpStream` and lets me remove all the manual buffer management for UDP codepaths, like the `UdpTrafficSocket` struct I had before.