Add "channels" forwarding mode
A common scenario when forwarding or proxying network traffic between two machines is when one initiates multiple connections to the same remote endpoint (IP:port combo). The sender expects to have two conceptual pipes to the other machine. Broker mode is not appropriate for this, because it broadcasts data between all known senders.

Channel mode creates groupings between two remote peers on different machines. When traffic arrives at one end of a channel grouping, it is forwarded to the other end of the channel grouping and not to any other connected sockets.

The user can enable channel mode using the new `--fm channels` or `-f c` for short. Broker mode has also been moved to the new forwarding mode arg with `--fm broker` or `-f b`. The default mode is `--fm none`, behavior unchanged from the previous default.

Now when enabling broker or channels mode, `max_clients` is automatically set to 10 but can be overridden by the user.

To make this work, I first had to introduce tracking of both the remote source and local destination (a new `RouteAddr` object) into `SourcedBytes` so that on every packet handling, I can look up the intended destination from a new `ChannelMap` object. Many places changed to use `RouteAddr`s instead of `SocketAddr`s.

I added a way to notify the router that a TcpStream has ended, so it can clear out the route right away. Without that, it had to wait for the next attempted send to the destination to encounter an error. This was also necessary to make reconnections for channels work properly: the reconnection would happen successfully but the old, broken route would still be tracked because no new traffic had tried sending out of it yet.