Add convenience for connecting to a target multiple times
For channel setup, you sometimes need multiple connections to the same target, to support the user's application that needs multiple connections. This change allows appending a multiplier to a target. For example, to connect to localhost on port 5000 twelve times, you can pass `localhost:5000x12`.

Also forgot to "cargo fmt" after the last change, so the diff is much larger than it should be. Argh.