Forward socket close in Channels mode; add Lingering Channels mode
When doing channels mode, and one socket disconnects, you often want that to be forwarded through to the other endpoint, fitting the mental model that netcrab is acting like a proxy in the middle. After all, programs base a lot of state around socket lifetime, and if someone reconnects to the channel without it being torn down, it might get confused.

So by default, channels mode, upon noticing a socket at one end of the channel closing, will now also disconnect the other end of the channel. If this behavior doesn't suit your needs, the old behavior to keep the other end of the channel open is available using `--fm lingering-channels` or `-f lc`.