Allow exiting the program after input reaches end of file
It can be convenient to pipe some input into netcrab and have it exit when the input has been sent. This adds a new option called `--exit-after-input` which does exactly that.

To achieve this, when EOF is reached, we send an empty payload from the local IO into the router, which signals to it to mark itself as done and close all sockets (when `--exit-after-input` is set). The caller of the router also has to check when the router is done, so it doesn't do things like try to reconnect the just-disconnected sockets or start listening for more connections.