Add throughput stats output and "fixed" input mode for perf testing
To test perf fixes, netcrab can now output stats every sec about the throughput sent and received. You can enable these stats using `--stats`.

To try to get a higher speed showing, it now supports a new input mode, `--i fixed` that creates a single buffer filled with random data and sends it repeatedly. The size of the buffer is controlled by `--sb` and the type of data in it is controlled by `--rvals`. This cuts out the relatively high time I found spent in random number generation.

Also to get a higher throughput, the user can now control the receive buffer using `--rb SIZE`.

Previously, the TCP send path was using a `TargetedBytes`, passing through the destination address unnecessarily to the socket sink. Not only was this sending more data to the sink than necessary, the sink had to use `SinkExt::with` to remove the destination address, because the socket sink accepts just a `Bytes`. That's a future executed on every send to the socket. This change removes that along with everything related to `TargetedBytes`, since it's not needed.

Internally, the `RandBytesIter` now doesn't depend on the whole `NcArgs` object and the logic of picking the random distribution is moved into the `RandBytesIter` class itself. To do this, I had to make the iterator it stores a `Box<dyn Iterator = u8>>` instead of the concrete templated type it had before. Maybe it wasn't totally necessary but it makes the code easier to deal with.

The throughput stats are stored in a new `StatsTracker` object and queried after servicing traffic to see if it needs to be printed.