Perf: remove mpsc channel when writing to TCP socket
Before, the TcpRouter would create an mpsc channel for each socket and hand back the output end for the socket to drive to the network. This additional layer caused sends that went too fast to pile up and consume more and more memory, as it outpaced the speed of the network stack.

Through the magic of `into_split()` we can split the TcpStream into an owned read and write half that can be sent to different places. Now the router can hold that object directly and remove the mpsc channel, instead writing directly to the socket.

With this, throughput on localhost increases by about 2x consistently.