Rename "broker" to "hub"; add convenience shortcut args; fix outbound connection bug
The name "broker" never felt right to me. It acts more like an oldschool Ethernet hub, so I renamed it everywhere to "hub".

Added short versions of a few args, in particular input modes and output modes, so you can do `-in -on` to silence input and output. Also `--eai` for `--exit-after-input`, which is a mouthful.

If max inbound connections isn't set, we will now set it to one *per listen address* specified by the user, basically the number of times the user gave `-l`/`-L`.

There was a bug where if any outbound connection to a target failed, the target was considered failed. The logic has been corrected to consider it failed only if *all* connections failed.

Minor change to defer creating and configuring the socket for an outbound connection until after a valid source address has been found.