Add support for executing external program as IO
This change allows netcrab to execute an external program in place of local IO using `--exec` or `-e`. The child process's stdin and stdout are connected to the network via mpsc channels, so remote clients can interact with the local program.

Along with this, I've changed `--sb` to focus only on setting the send buffer socket option. It no longer controls the datagram size or size of the buffer passed to send(). That is now controlled by `--ss`.