Remove mpsc channel between local input and router
Before this change, to read from stdin or the terminal, the program would spin up a thread and send local input through an mpsc channel. Then something would do a `send_all` to drive the input into the router.

But since the router's sink can be cloned, it could be possible to pass this directly to the stdin reader thread and bypass the CPU and memory cost of the channel.

To make this change, all the places that were previously referencing a `LocalInputStream` need to change to use a future that exits when the local input is done. Internally the future will drive input to the router sink. This allows the stdin reader thread to call `router_sink.send()` and other input modes like random generation to call `router_sink.send_all()`.

At program exit it's possible that the stdin reader task blocks on reading from input, so we have to shut down the Tokio runtime rudely to kill the task.

Also fix a bug where the result of the last outbound connection wasn't being surfaced as the program's exit code.

Also remove an incorrect assert in the codepaths the router send path and the socket close path notice the socket closure and both call `cleanup_route`.

Also update to latest crate versions.