Use `BytesMut::freeze()` rather than `Bytes::copy_from_slice` where possible
In a few places, I had a pattern where I'd have some local buffer that I filled with data. When I was done with it, I'd call `Bytes::copy_from_slice`. A slightly more efficient way is to call `BytesMut::with_capacity` and fill directly into that buffer. When I'm done, I can call `freeze()` to turn it into a `Bytes`.

This eliminates a buffer copy in `copy_from_slice`. I was able to apply this technique in a few places and gain, uh, negligible throughput improvements.