Use tokio `ChildStdin` instead of setting up own writer thread
Previously we used a separate thread and an mpsc channel to write to a child process's stdin, since is it a `Write` but we need an `AsyncWrite` to work properly with the rest of our async code.

I just discovered `ChildStdin`, which takes care of this for me, so I can remove the separate thread and channel entirely.