Reduce the size of per-packet tracking info
Every packet that arrives from the network has to get tagged with its origin so the router can decide where to forward it to, which channels need to be created, etc.. Previously, this was a `RouteAddr`, which is a local and remote `SocketAddr`, totaling 64 bytes!

That means for a 1-byte incoming packet, we have to use 65 bytes of space in the mpsc channel and copy that around. That feels like a lot.

In this change, the `RouteAddr` is changed into a `RouteId`, which is just a `u16` wrapper. Most of the code that was previously using `RouteAddr`s for decision making purposes (comparing, looking up in maps, etc.) has now been changed to use `RouteId`s instead.

`RouteAddr`s are still used for user-visible logging and for a little bit of decision making in UDP paths and channels (channels are only created between two different IPs), so a new object the `RouteDb` is introduced to track which routes are known and to store the mapping between them and route ids.

The use of the `RouteDb` is pretty straightforward for TCP sockets, but for UDP, every datagram comes with the remote peer, so the db needs to be consulted in every UDP socket's reading task. It ends up having to be wrapped as an `Arc<RwLock<RouteDb>>`.

Theoretically it should have been faster to use this small data size everywhere, but I wasn't able to actually measure anything. Oh well.

Also fixed a small bug where channels weren't disconnecting the other end in UDP. Though I'm not sure if that codepath can ever be hit in practice due to the way UDP failures work.