Use functions instead of closures when possible
There are a number of calls, especially to `map`, that take a closure with a single function call on the argument. This change swaps them for plain function calls instead, which should have a slight reduction in code size.

For example:

```rust
map(|e| e.clone)
```

is replaced by

```rust
map(std::clone::Clone::clone)
```