Add support for PlayFab Quality of Service beacon client
netcrab can now act as a client for [PlayFab Quality of Service beacons](https://learn.microsoft.com/en-us/gaming/playfab/features/multiplayer/servers/using-qos-beacons-to-measure-player-latency-to-azure) if you specify `-i pfqoscli`. This allows for some simple latency testing. When in this mode, netcrab prints out the latest round trip time to the beacon periodically.

The ping period and its output (and now also the period for printing out throughput stats) is controlled by a new argument, `--stats-time`. The latest RTT will only be measured if the response from the beacon server is validated according to the documented protocol above.

The payloads sent to the beacon server consist of a small fixed-size header followed by random data. The random portion is controlled by `RandConfig`, so the same arguments that control `--i rand`.

Because `RandBytesIter` now sends a portion as as a fixed buffer, I removed `FixedBytesIter` in favor of a mode in `RandBytesIter` where it doesn't engage the random number generation at all if `RandConfig::size_max` is zero.

The latency stats are stored in the `RouteDb` alongside the route address in a new struct called `RouteInfo`. This could be used to store other per-route tracking in the future.

Because the latency is tracked on a per-route basis, you can measure against multiple QoS servers at a time.