Add support for PlayFab Quality of Service beacon server
netcrab can now act as a [PlayFab Quality of Service beacon](https://learn.microsoft.com/en-us/gaming/playfab/features/multiplayer/servers/using-qos-beacons-to-measure-player-latency-to-azure) server if you specify `-i pfqossrv`. This can be used in conjunction with another instance using `-i pfqoscli` to do simple latency testing.

This mode operates very similarly to `-i echo` except that it examines the incoming data, and if it starts with 0xff 0xff as documented in the link above, it inverts those two bytes to 0 0 when sending the echo back.